Rename the facade module to client.
